warmup validator status in exporter
